src/backend/requirements.txt
#____Required Django packages_____
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
PyMySQL==1.1.0
cryptography==42.0.3

#____Required Pydantic packages_____
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1
